requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
multidict==6.2.0
propcache==0.3.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
SQLAlchemy==2.0.39
typing_extensions==4.13.0
yarl==1.18.3
apscheduler==3.10.1
psycopg2-binary==2.9.9